requirements.txt
fastapi==0.110.0
certifi==2024.7.4
uvicorn[standard]==0.29.0
pre-commit==2.21.0
redis[hiredis]==5.0.3
rq==1.16.1
python-dateutil==2.8.2
